Get rid of build number
