+++ .github/workflows/main.yml
@@ -241,14 +241,13 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'Stateless tests (debug, actions)'
           REPO_COPY: ${{runner.temp}}/stateless_debug/ClickHouse
-          REQUIRED_BUILD_NUMBER: 7
           KILL_TIMEOUT: 10800
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 functional_test_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER $KILL_TIMEOUT
+          python3 functional_test_check.py "$CHECK_NAME" $KILL_TIMEOUT
       - name: Cleanup
         if: always()
         run: |
@@ -271,14 +270,13 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'Stateful tests (debug, actions)'
           REPO_COPY: ${{runner.temp}}/stateful_debug/ClickHouse
-          REQUIRED_BUILD_NUMBER: 7
           KILL_TIMEOUT: 3600
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 functional_test_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER $KILL_TIMEOUT
+          python3 functional_test_check.py "$CHECK_NAME" $KILL_TIMEOUT
       - name: Cleanup
         if: always()
         run: |
@@ -301,14 +299,13 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'Stateless tests flaky check (address, actions)'
           REPO_COPY: ${{runner.temp}}/stateless_flaky_asan/ClickHouse
-          REQUIRED_BUILD_NUMBER: 3
           KILL_TIMEOUT: 3600
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 functional_test_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER $KILL_TIMEOUT
+          python3 functional_test_check.py "$CHECK_NAME" $KILL_TIMEOUT
       - name: Cleanup
         if: always()
         run: |
@@ -331,13 +328,12 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'Stress tests (debug, actions)'
           REPO_COPY: ${{runner.temp}}/stress_debug/ClickHouse
-          REQUIRED_BUILD_NUMBER: 7
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 stress_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER
+          python3 stress_check.py "$CHECK_NAME"
       - name: Cleanup
         if: always()
         run: |
@@ -360,13 +356,12 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'AST fuzzer (debug, actions)'
           REPO_COPY: ${{runner.temp}}/ast_fuzzer_debug/ClickHouse
-          REQUIRED_BUILD_NUMBER: 7
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 ast_fuzzer_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER
+          python3 ast_fuzzer_check.py "$CHECK_NAME"
       - name: Cleanup
         if: always()
         run: |
@@ -389,13 +384,12 @@ jobs:
 #          REPORTS_PATH: ${{runner.temp}}/reports_dir
 #          CHECK_NAME: 'Integration tests (asan, actions)'
 #          REPO_COPY: ${{runner.temp}}/integration_tests_asan/ClickHouse
-#          REQUIRED_BUILD_NUMBER: 3
 #        run: |
 #          sudo rm -fr $TEMP_PATH
 #          mkdir -p $TEMP_PATH
 #          cp -r $GITHUB_WORKSPACE $TEMP_PATH
 #          cd $REPO_COPY/tests/ci
-#          python3 integration_test_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER
+#          python3 integration_test_check.py "$CHECK_NAME"
 #      - name: Cleanup
 #        if: always()
 #        run: |
@@ -418,13 +412,12 @@ jobs:
           REPORTS_PATH: ${{runner.temp}}/reports_dir
           CHECK_NAME: 'Unit tests (asan, actions)'
           REPO_COPY: ${{runner.temp}}/unit_tests_asan/ClickHouse
-          REQUIRED_BUILD_NUMBER: 3
         run: |
           sudo rm -fr $TEMP_PATH
           mkdir -p $TEMP_PATH
           cp -r $GITHUB_WORKSPACE $TEMP_PATH
           cd $REPO_COPY/tests/ci
-          python3 unit_tests_check.py "$CHECK_NAME" $REQUIRED_BUILD_NUMBER
+          python3 unit_tests_check.py "$CHECK_NAME"
       - name: Cleanup
         if: always()
         run: |
